refactor: Create a separate module for the config load phase.
The mapping from ConfigModel to more specific beans is moved from the factory module to the
configuration module.
