+++ src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationFactory.java
@@ -6,6 +6,7 @@ import javax.inject.Singleton;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripSchedule;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.standalone.config.ConfigModel;
+import org.opentripplanner.standalone.config.configure.ConfigModule;
 import org.opentripplanner.transit.raptor.configure.RaptorConfig;
 import org.opentripplanner.visualizer.GraphVisualizer;
 
@@ -13,7 +14,7 @@ import org.opentripplanner.visualizer.GraphVisualizer;
  * Dagger dependency injection Factory to create components for the OTP construct application phase.
  */
 @Singleton
-@Component(modules = ConstructApplicationModule.class)
+@Component(modules = ConfigModule.class)
 public interface ConstructApplicationFactory {
   RaptorConfig<TripSchedule> raptorConfig();
